+++ referra-api/src/main/java/com/sbproject/referra/model/ReferralStatus.java
@@ -1,0 +1,8 @@
+package com.sbproject.referra.model;
+
+public enum ReferralStatus {
+    OPEN,
+    LIMITED,
+    PAUSED,
+    CLOSED
+}