+++ referra-api/src/test/java/com/sbproject/referra/service/CompanyServiceTest.java
@@ -23,6 +23,7 @@ class CompanyServiceTest {
     public static final String TEST_COMPANY_NAME = "TestCo";
     public static final String TEST_DOMAIN = "testco.com";
     public static final String TEST_COMPANY_URL = "https://testco.com";
+
     @Mock
     private CompanyRepository companyRepository;